Travis tests for local mashine

+++ travis/test.sh
@@ -39,17 +39,16 @@ go list -f 'go test -v -tags integration -race -covermode atomic -coverprofile {
 # Merge coverage profiles.
 COVERAGE_FILES=`ls -1 *.coverprofile 2>/dev/null | wc -l`
 if [ $COVERAGE_FILES != 0 ]; then
-    gocovmerge `ls *.coverprofile` > coverage.txt
-    rm *.coverprofile
+	# check program `gocovmerge` is exist
+	if which gocovmerge >/dev/null 2>&1; then
+		gocovmerge `ls *.coverprofile` > coverage.txt
+		rm *.coverprofile
+	fi
 fi
 
 # Print stats
 echo "Unit tests: " $(grep "=== RUN" $OUTFILE | wc -l)
 echo "Integration tests: " $(grep "# Total tests" $OUTFILE | cut -c21-)
-
-# Remove the outfile so it is not printed when an error happens beyond this
-# point.
-rm $OUTFILE
 
 # These steps are from the README to verify it can be installed and run as
 # documented.
@@ -59,20 +58,36 @@ export C2GO_DIR=$GOPATH/src/github.com/elliotchance/c2go
 export C2GO=$C2GO_DIR/c2go
 
 echo "Run: c2go transpile prime.c"
-$C2GO transpile $C2GO_DIR/examples/prime.c
-echo "47" | go run prime.go
+$C2GO transpile -o=/tmp/prime.go $C2GO_DIR/examples/prime.c
+echo "47" | go run /tmp/prime.go
 if [ $($C2GO -v | wc -l) -ne 1 ]; then exit 1; fi
-if [ $(cat prime.go | wc -l) -eq 0 ]; then exit 1; fi
+if [ $(cat /tmp/prime.go | wc -l) -eq 0 ]; then exit 1; fi
 if [ $($C2GO ast $C2GO_DIR/examples/prime.c | wc -l) -eq 0 ]; then exit 1; fi
 
 # This will have to be updated every so often to the latest version. You can
 # find the latest version here: https://sqlite.org/download.html
 export SQLITE3_FILE=sqlite-amalgamation-3190300
 
+# Variable for location of temp sqlite files
+SQLITE_TEMP_FOLDER="/tmp/SQLITE"
+
+# Remove temp folder of SQLITE3
+if [ -d $SQLITE_TEMP_FOLDER ]; then
+	rm -r $SQLITE_TEMP_FOLDER
+fi
+
+# Create a folder SQLITE
+mkdir $SQLITE_TEMP_FOLDER
+
 # Download Sqlite3 amalgamated source.
-curl https://sqlite.org/2017/$SQLITE3_FILE.zip > /tmp/$SQLITE3_FILE.zip
-unzip /tmp/$SQLITE3_FILE.zip -d /tmp
+curl https://sqlite.org/2017/$SQLITE3_FILE.zip > $SQLITE_TEMP_FOLDER/$SQLITE3_FILE.zip
+unzip $SQLITE_TEMP_FOLDER/$SQLITE3_FILE.zip -d $SQLITE_TEMP_FOLDER
 
 # Transpile the SQLite3 files.
-./c2go transpile /tmp/sqlite-amalgamation-3190300/shell.c
-./c2go transpile /tmp/sqlite-amalgamation-3190300/sqlite3.c
+echo "Transpiling shell.c..."
+./c2go transpile -o=$SQLITE_TEMP_FOLDER/shell.go   $SQLITE_TEMP_FOLDER/sqlite-amalgamation-3190300/shell.c
+echo "Transpiling sqlite3.c..."
+./c2go transpile -o=$SQLITE_TEMP_FOLDER/sqlite3.go $SQLITE_TEMP_FOLDER/sqlite-amalgamation-3190300/sqlite3.c
+
+# Remove c2go executable file
+rm ./c2go
